docker_version/odin-libraries/python/requirements.txt
pymongo==3.10.1
requests==2.23.0
ruamel.yaml==0.16.10 
